+++ tests/unit/memory_storage/resource_clients/test_request_queue.py
@@ -108,6 +108,7 @@ async def test_add_record(request_queue_client: RequestQueueClient) -> None:
     rq_info = await request_queue_client.get()
     assert rq_info is not None
     assert rq_info['pendingRequestCount'] == rq_info['totalRequestCount'] == 2
+    assert rq_info['handledRequestCount'] == 0
 
 
 async def test_get_record(request_queue_client: RequestQueueClient) -> None:
@@ -135,28 +136,46 @@ async def test_update_record(request_queue_client: RequestQueueClient) -> None:
     rq_info_before_update = await request_queue_client.get()
     assert rq_info_before_update is not None
     assert rq_info_before_update['pendingRequestCount'] == 1
+    assert rq_info_before_update['handledRequestCount'] == 0
     request_update_info = await request_queue_client.update_request({**request, 'handledAt': datetime.now(timezone.utc)})
     assert request_update_info['wasAlreadyHandled'] is False
     rq_info_after_update = await request_queue_client.get()
     assert rq_info_after_update is not None
     assert rq_info_after_update['pendingRequestCount'] == 0
+    assert rq_info_after_update['handledRequestCount'] == 1
 
 
 async def test_delete_record(request_queue_client: RequestQueueClient) -> None:
     request_url = 'https://apify.com'
-    request_info = await request_queue_client.add_request({
-        'uniqueKey': request_url,
+    pending_request_info = await request_queue_client.add_request({
+        'uniqueKey': 'pending',
         'url': request_url,
     })
-    rq_info_before_update = await request_queue_client.get()
-    assert rq_info_before_update is not None
-    assert rq_info_before_update['pendingRequestCount'] == 1
-    await request_queue_client.delete_request(request_info['requestId'])
-    rq_info_after_update = await request_queue_client.get()
-    assert rq_info_after_update is not None
-    assert rq_info_after_update['pendingRequestCount'] == 0
+    handled_request_info = await request_queue_client.add_request({
+        'uniqueKey': 'handled',
+        'url': request_url,
+        'handledAt': datetime.now(tz=timezone.utc),
+    })
+
+    rq_info_before_delete = await request_queue_client.get()
+    assert rq_info_before_delete is not None
+    assert rq_info_before_delete['pendingRequestCount'] == 1
+    assert rq_info_before_delete['pendingRequestCount'] == 1
+
+    await request_queue_client.delete_request(pending_request_info['requestId'])
+    rq_info_after_first_delete = await request_queue_client.get()
+    assert rq_info_after_first_delete is not None
+    assert rq_info_after_first_delete['pendingRequestCount'] == 0
+    assert rq_info_after_first_delete['handledRequestCount'] == 1
+
+    await request_queue_client.delete_request(handled_request_info['requestId'])
+    rq_info_after_second_delete = await request_queue_client.get()
+    assert rq_info_after_second_delete is not None
+    assert rq_info_after_second_delete['pendingRequestCount'] == 0
+    assert rq_info_after_second_delete['handledRequestCount'] == 0
+
     # Does not crash when called again
-    await request_queue_client.delete_request(request_info['requestId'])
+    await request_queue_client.delete_request(pending_request_info['requestId'])
 
 
 async def test_forefront(request_queue_client: RequestQueueClient) -> None: