+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 - fixed parsing messages from the platform events websocket when they have no event data
 - fixed `EventManager` not waiting for platform events websocket connection during initialization
 - fixed local `RequestQueueClient` not respecting the `forefront` argument
+- fixed local `RequestQueueClient` not counting the `handledRequestCount` property
 
 ### Internal changes